Move cycle state management to breathingmanager

+++ src/breath_cycle.rs
@@ -1,8 +1,9 @@
+#[derive(Clone,Copy,PartialEq)]
 pub struct BreathCycle{
-    inhale_duration : f64,
-    break1_duration : f64,
-    exhale_duration : f64,
-    break2_duration : f64,
+    inhale_duration : i16,
+    break1_duration : i16,
+    exhale_duration : i16,
+    break2_duration : i16,
 }
 
 #[derive(Clone,Copy,PartialEq)]
@@ -31,7 +32,7 @@ impl CycleState{
 }
 
 impl BreathCycle{
-    pub fn new(inhale_duration : f64, break1_duration : f64, exhale_duration : f64, break2_duration : f64) -> Self{
+    pub fn new(inhale_duration : i16, break1_duration : i16, exhale_duration : i16, break2_duration : i16) -> Self{
         BreathCycle{
             inhale_duration,
             break1_duration,
@@ -42,34 +43,34 @@ impl BreathCycle{
 
     
     //Setters
-    pub fn set_inhale_duration(&mut self, duration : f64){
+    pub fn set_inhale_duration(&mut self, duration : i16){
         self.inhale_duration = duration;
     }
-    pub fn set_break1_duration(&mut self, duration : f64){
+    pub fn set_break1_duration(&mut self, duration : i16){
         self.break1_duration = duration;
     }
-    pub fn set_exhale_duration(&mut self, duration : f64){
+    pub fn set_exhale_duration(&mut self, duration : i16){
         self.exhale_duration = duration;
     }
-    pub fn set_break2_duration(&mut self, duration : f64){
+    pub fn set_break2_duration(&mut self, duration : i16){
         self.break2_duration = duration;
     }
     
     //Getters
-    pub fn total_cycle_duration(&self) -> f64{
+    pub fn total_cycle_duration(&self) -> i16{
         self.inhale_duration + self.break1_duration + self.exhale_duration + self.break2_duration
     }
-    pub fn inhale_duration(&self) -> f64 {self.inhale_duration}
-    pub fn break1_duration(&self) -> f64 {self.break1_duration}
-    pub fn exhale_duration(&self) -> f64 {self.exhale_duration}
-    pub fn break2_duration(&self) -> f64 {self.break2_duration}
-    pub fn get_state_duration(& self, cycle_state: &CycleState) -> f64{
+    pub fn inhale_duration(&self) -> i16 {self.inhale_duration}
+    pub fn break1_duration(&self) -> i16 {self.break1_duration}
+    pub fn exhale_duration(&self) -> i16 {self.exhale_duration}
+    pub fn break2_duration(&self) -> i16 {self.break2_duration}
+    pub fn get_state_duration(& self, cycle_state: &CycleState) -> i16{
         match cycle_state {
             CycleState::Inhale => self.inhale_duration,
             CycleState::Break1 => self.break1_duration,
             CycleState::Exhale => self.exhale_duration,
             CycleState::Break2 => self.break2_duration,
-            CycleState::None => 0.0,
+            CycleState::None => 0,
         }
     }
 }
